SDL: Fix live edit when start path contains unicode characters #123
